Organized server files into ControllerServer and InstructionServer folders
