+++ start.py
@@ -1,0 +1,4 @@
+from InstructionsServer.instructions import startInstructionServer
+from ControllerServer.server import startControllerServer
+
+
